requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==1.10.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.8.2
blis==0.7.9
boto3==1.26.115
botocore==1.29.115
bpemb==0.3.4
breadability==0.1.20
catalogue==2.0.8
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
cleantext==1.1.4
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
compressed-rtf==1.0.6
confection==0.0.4
conllu==4.5.2
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
cymem==2.0.7
datasets==2.11.0
Deprecated==1.2.13
dill==0.3.6
Django==4.1.7
django-bootstrap5==22.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-thread==0.0.1
djangorestframework==3.14.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
eng-to-ipa==0.0.2
exceptiongroup==1.1.0
extract-msg==0.28.7
filelock==3.11.0
flair==0.12.2
flake8==6.0.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
fuzzywuzzy==0.18.0
gdown==4.4.0
gensim==4.3.1
huggingface-hub==0.13.4
hyperopt==0.2.7
idna==3.4
IMAPClient==2.1.0
iniconfig==2.0.0
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
Levenshtein==0.21.0
lit==16.0.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
packaging==23.0
pandas==2.0.0
pathy==0.10.1
pdfminer.six==20191110
Pillow==9.5.0
pluggy==1.0.0
pptree==3.1
preshed==3.0.8
protobuf==3.20.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
py4j==0.10.9.7
pyarrow==11.0.0
pycodestyle==2.10.0
pycountry==22.3.5
pycryptodome==3.17
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-Levenshtein==0.21.0
python-pptx==0.6.21
pytorch_revgrad==0.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentencepiece==0.1.98
six==1.12.0
sklearn==0.0.post4
smart-open==6.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
sqlitedict==2.1.0
sqlparse==0.4.3
srsly==2.4.6
sumy==0.11.0
sympy==1.11.1
tabulate==0.9.0
tesseract==0.1.3
textblob==0.17.1
textract==1.6.5
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.3
transformers==4.28.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
vaderSentiment==3.3.2
wasabi==1.1.1
wcwidth==0.2.6
Wikipedia-API==0.5.8
wordnik==2.1.3
wrapt==1.15.0
xlrd==1.2.0
XlsxWriter==3.1.0
xxhash==3.2.0
yarl==1.8.2
